backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pytesseract>=0.3.10
Pillow>=10.2.0
pandas>=2.1.3
openpyxl>=3.0.0>=3.1.2
selenium>=4.15.2
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
pdf2image>=1.16.3
pypdf2>=3.0.1
PyMuPDF>=1.23.0
numpy>=1.26.2
opencv-python>=4.8.1.78

